Revert 'simplified swift templates error handling'
69c5ec98c82070682a4dce5839250cd89bfb9e4a

+++ SourcerySwift/Sources/SwiftTemplate.swift
@@ -15,6 +15,11 @@ private enum Delimiters {
     static let close = "%>"
 }
 
+private struct ProcessResult {
+    let output: String
+    let error: String
+}
+
 open class SwiftTemplate {
 
     public let sourcePath: Path
@@ -189,13 +194,12 @@ open class SwiftTemplate {
         let data = NSKeyedArchiver.archivedData(withRootObject: context)
         try serializedContextPath.write(data)
 
-        do {
-            let result = try Process.runCommand(path: binaryPath.description,
-                                                arguments: [serializedContextPath.description])
-            return result
-        } catch let error as SwiftTemplateError {
-            throw "\(sourcePath): \(error.reason)"
+        let result = try Process.runCommand(path: binaryPath.description,
+                                            arguments: [serializedContextPath.description])
+        if !result.error.isEmpty {
+            throw "\(sourcePath): \(result.error)"
         }
+        return result.output
     }
 
     func build() throws -> Path {
@@ -217,14 +221,20 @@ open class SwiftTemplate {
                 "-Xlinker", "-headerpad_max_install_names"
         ]
 
-        try Process.runCommand(path: "/usr/bin/swiftc",
-                               arguments: arguments)
+        let compilationResult = try Process.runCommand(path: "/usr/bin/swiftc",
+                                                       arguments: arguments)
+        if !compilationResult.error.isEmpty {
+            throw compilationResult.error
+        }
         
-        try Process.runCommand(path: "/usr/bin/install_name_tool",
-                               arguments: [
-                                "-add_rpath",
-                                "@executable_path/../",
-                                binaryFile.description])
+        let linkingResult = try Process.runCommand(path: "/usr/bin/install_name_tool",
+                                                   arguments: [
+                                                    "-add_rpath",
+                                                    "@executable_path/../",
+                                                    binaryFile.description])
+        if !linkingResult.error.isEmpty {
+            throw linkingResult.error
+        }
 
         try? mainFile.delete()
 
@@ -234,12 +244,16 @@ open class SwiftTemplate {
     private func copyFramework(to path: Path) throws {
         let sourceryFramework = SwiftTemplate.frameworksPath + "SourceryRuntime.framework"
 
-        try Process.runCommand(path: "/usr/bin/rsync", arguments: [
+        let copyFramework = try Process.runCommand(path: "/usr/bin/rsync", arguments: [
             "-av",
             "--force",
             sourceryFramework.description,
             path.description
             ])
+
+        if !copyFramework.error.isEmpty {
+            throw copyFramework.error
+        }
     }
 
 }
@@ -250,14 +264,6 @@ fileprivate extension SwiftTemplate {
         return Path(Bundle(for: SwiftTemplate.self).bundlePath +  "/Versions/Current/Frameworks")
     }
 
-}
-
-struct SwiftTemplateError: Error, CustomStringConvertible {
-    let reason: String
-    
-    var description: String {
-        return reason
-    }
 }
 
 // swiftlint:disable:next force_try
@@ -276,8 +282,7 @@ private extension String {
 }
 
 private extension Process {
-    @discardableResult
-    static func runCommand(path: String, arguments: [String], environment: [String: String] = [:]) throws -> String {
+    static func runCommand(path: String, arguments: [String], environment: [String: String] = [:]) throws -> ProcessResult {
         let task = Process()
         task.launchPath = path
         task.arguments = arguments
@@ -303,11 +308,7 @@ private extension Process {
         let output = String(data: outputData, encoding: .utf8) ?? ""
         let error = String(data: errorData, encoding: .utf8) ?? ""
 
-        if !errorData.isEmpty {
-            throw SwiftTemplateError(reason: error)
-        }
-
-        return output
+        return ProcessResult(output: output, error: error)
     }
 }
